requirements.txt
asgiref==3.3.1
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
pytz==2020.4
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.2
